Add Rake task to test against ActiveRecord test suite
