+++ .gitignore
@@ -8,3 +8,5 @@ Gemfile.lock
 /travis/*.lock
 /test/database_local.yml
 .idea
+debug.log
+test/db/*
