+++ CONTRIBUTING.md
@@ -37,6 +37,22 @@ Make sure the tests pass:
 
 `bundle exec rake`
 
+To run a specific test, use the `POSTGIS_TEST_FILES` environment variable:
+
+`POSTGIS_TEST_FILES=test/cases/ddl_test.rb bundle exec rake`
+
+If you are testing a feature against the ActiveRecord test suite run:
+
+`bundle exec rake test:activerecord`
+
+Files can be specified with the `AR_TEST_FILES` environment variable:
+
+`AR_TEST_FILES test/cases/adapters/postgresql/*_test.rb bundle exec rake test:activerecord`
+
+To test with both local and ActiveRecord tests, run:
+
+`bundle exec rake test:all`
+
 Run tests against the test gemfiles:
 
 run `rake appraisal` or run the tests manually:
